long double

+++ Ymir/src/types.h
@@ -86,8 +86,8 @@ namespace ymir {
     *
     * \brief Type of stored probabilities of different events.
     */
-    typedef double prob_t;
-//    typedef long double prob_t;
+//    typedef double prob_t;
+    typedef long double prob_t;
     #endif
